+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Bugfix: Bearbeiten von Herstellern invalidiert Objektcache für Artikel nicht
 * Bugfix: Newsbeiträge aus deaktivierten Kategorien werden auf der Startseite angezeigt
 * Bugfix: Globale Variable $AktuelleSeite zeigt bei Newsdetails falschen Typ an
+* Verbesserung: Neuer Hook 220 HOOK_GET_NEWS in gibNews()
 
 ## [4.03.1]
 * Bugfix: Sprachwechsel in einigen Linkgruppen unvollständig